+++ packages/nuxt/src/module.ts
@@ -10,6 +10,7 @@ import { consoleSandbox } from '@sentry/core';
 import * as path from 'path';
 import type { SentryNuxtModuleOptions } from './common/types';
 import { addDynamicImportEntryFileWrapper, addSentryTopImport, addServerConfigToBuild } from './vite/addServerConfig';
+import { addMiddlewareImports, addMiddlewareInstrumentation } from './vite/middlewareConfig';
 import { setupSourceMaps } from './vite/sourceMaps';
 import { addOTelCommonJSImportAlias, findDefaultSdkInitFile } from './vite/utils';
 
@@ -110,7 +111,16 @@ export default defineNuxtModule<ModuleOptions>({
       };
     });
 
+    // Preps the the middleware instrumentation module.
+    if (serverConfigFile) {
+      addMiddlewareImports();
+    }
+
     nuxt.hooks.hook('nitro:init', nitro => {
+      if (serverConfigFile) {
+        addMiddlewareInstrumentation(nitro);
+      }
+
       if (serverConfigFile?.includes('.server.config')) {
         consoleSandbox(() => {
           const serverDir = nitro.options.output.serverDir;